implement selection of lines, circles and polygons.

+++ trunk/engine/vg/vg_ellipse.c
@@ -1,4 +1,4 @@
-/*	$Csoft: vg_ellipse.c,v 1.13 2005/05/21 03:32:55 vedge Exp $	*/
+/*	$Csoft: vg_ellipse.c,v 1.14 2005/06/04 04:48:44 vedge Exp $	*/
 
 /*
  * Copyright (c) 2004, 2005 CubeSoft Communications, Inc.
@@ -79,10 +79,10 @@ extent(struct vg *vg, struct vg_element *vge, struct vg_rect *r)
 	r->h = vge->vg_arc.h;
 }
 
-static int
-intsect(struct vg *vg, struct vg_element *vge, int x, int y)
+static float
+intsect(struct vg *vg, struct vg_element *vge, double x, double y)
 {
-	return (INT_MAX);
+	return (FLT_MAX);
 }
 
 const struct vg_element_ops vg_ellipse_ops = {
